Cull large builder boxes because they don't work yet (#3139)
The very large builder boxes don't fit in either width or weight for a lot of vehicles that they'll load into. Can be fixed later, but not in 3.5. Also the gameplay need for them is questionable: We probably don't want people building or repairing that much.

Buffed the "large" box to 1500 as a placeholder.